backend/requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
pocketbase==0.10.1
python-dotenv==1.0.1
requests==2.31.0
selectolax==0.3.21
sniffio==1.3.1
typing_extensions==4.11.0
urllib3==2.2.1
